feat(6): check correctness function (via window)
The user can click on the button to check the correctness of their sudoku puzzle solution.

When the button is clicked on, a message is displayed on a new window. The user can close this window to continue playing the sudoku game.

There are 3 different possible messages, depending on the correctness of the solution:

(1) Sudoku puzzle unfinished: 'You have not finished yet!'
(2) Sudoku puzzle solved incorrectly: 'Try again!'
(3) Sudoku puzzle solved successfully: 'It is correct!'